+++ config/constants.php
@@ -1,0 +1,7 @@
+<?php
+
+// for user privilege
+if (!defined('FORBIDDEN')) define('FORBIDDEN', 0);
+if (!defined('ONLY_SEE')) define('ONLY_SEE', 1);
+if (!defined('CAN_CRUD')) define('CAN_CRUD', 2);
+if (!defined('ALL_ACCESS')) define('ALL_ACCESS', 3);